BOGUE: util.serveur: servir() ne prenait pas en compte les liens symboliques
# Lorsqu'/usr/local/etc/rc.d/truc était un lien symbolique vers /usr/local/truc-x.y.z/etc/rc.d/truc, il n'était pas pris en compte comme serveur lançable.

darcs-hash:7c3018818e315d283c9b6d29e56d4b49f60dd933

+++ util.serveur.sh
@@ -64,7 +64,7 @@ servir()
 			local racine
 			for racine in $dest /usr/local ""
 			do
-				[ -f "$racine/$initds/$serveur" ] || continue # On ne teste pas en -x, car on n'est pas forcément root, donc s'il est exécutable simplement par root le -x renverra faux.
+				[ -f "$racine/$initds/$serveur" -o -L "$racine/$initds/$serveur" ] || continue # On ne teste pas en -x, car on n'est pas forcément root, donc s'il est exécutable simplement par root le -x renverra faux.
 				case "$action" in
 					remove)
 						sudoku -f rm -f "$dest/$initds/$serveur" `find "$racine"/etc/rc[0-9].d/ -type f 2> /dev/null | grep "/[SK][0-9]*$serveur$"`
